+++ packages/cli/src/controller/migrate/index.ts
@@ -1,0 +1,8 @@
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+export * from './migrate-controller';
+export * from './manifest/migrate-manifest.controller';
+export * from './abis/migrate-abis.controller';
+export * from './schema/migrate-schema.controller';
+export * from './constants';