+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Migrate a subgraph project to subquery project
+  - enble migrate its `subgraph.yaml` to `project.ts`
+
 ## [4.6.0] - 2024-04-24
 ### Added
 - support env & env commands (#2345)